+++ CHANGELOG.md
@@ -14,8 +14,12 @@ This is used for recording Invader's changes. This changelog is based on
   tag
   
 ### Fixed
+- invader-archive: Fixed .model references being converted to .gbxmodel when
+  checking for dependencies
 - invader-build: Fixed some scenery spawns warning about fullbright when it is
   not an issue
+- invader-dependency: Fixed .model references being converted to .gbxmodel when
+  checking for dependencies
 
 ## [0.37.1] - 2020-09-15
 ### Added